% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDupsCNA.R
\name{addDupsCNA}
\alias{addDupsCNA}
\title{A function to match trios with CNA data using the package "org.Hs.eg.db"}
\usage{
addDupsCNA(final, cna)
}
\arguments{
\item{final}{A trios data matrix with 4 columns; gene name, methylation row, cna row, gene expression row}

\item{cna}{Copy Number Alteration dataset; genes in rows, individuals in columns}
}
\value{
a list with two data matrix

final

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The trios data matrix with NAs in the CNA column replaced with row number from the CNA dataset
}\if{html}{\out{</div>}}

dup.final

\if{html}{\out{<div class="sourceCode">}}\preformatted{ A data matrix with trios for genes that are matched with multiple entrez IDs in the CNA data
}\if{html}{\out{</div>}}
}
\description{
After matching the trios based on gene names and entrez ID, we still have missing data in the trios data matrix so we match the genes with a package available with additional information on the gene.
}
\examples{
#use the function to match trios using gene name and entrez ID
trios.df = findTrioAll(meth, cna, gene, 5, 3, 3)
trios.df

#use the function match additional entries in the CNA column of trios data using the package "org.Hs.eg.db"
result = addDupsCNA(trios.df, cna)

#initial trios data with entries in the CNA column filled in after matching
result[[1]]

#additional trios data for genes that were matched and had multiple entrez id matches in the CNA data
result[[2]]

}
\seealso{
\code{\link[=findDups]{findDups()}} to find duplicated rows
}
