% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addDupsGENE.R
\name{addDupsGENE}
\alias{addDupsGENE}
\title{A function to match trios with Gene Expression data using the package "org.Hs.eg.db"}
\usage{
addDupsGENE(final, dup.final, gene)
}
\arguments{
\item{final}{A trios data matrix with 4 columns; gene name, methylation row, cna row, gene expression row}

\item{dup.final}{A data matrix with trios for genes that are matched with multiple entrez IDs; gene name, methylation row, cna row, gene expression row}

\item{cna}{Gene Expression dataset; genes in rows, individuals in columns}
}
\value{
a list with two data matrix

final.res

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The trios data matrix with both final and dup.final data merged with NAs replaced by row numbers in Gene Expression column
}\if{html}{\out{</div>}}
}
\description{
After matching the trios based on gene names and entrez ID, we still have missing data in the trios data matrix so we match the genes with a package available with additional information on the gene.
}
\examples{
#use the function to match trios using gene name and entrez ID
trios.df = findTrioAll(meth, cna, gene, 5, 3, 3)
trios.df

#use the function to match additional entries in the CNA column of trios data using the package "org.Hs.eg.db"
result = addDupsCNA(trios.df, cna)

#initial trios data with entries in the CNA column filled in after matching
result[[1]]

#additional trios data for genes that were matched and had multiple entrez id matches in the CNA data
result[[2]]

#use the function to match additional entries in the Gene Expression column of trios data using the package "org.Hs.eg.db"
#It also merges the intial and additional trios data (res[[1]] and res[[2]]) and returns one data matrix
final.trios.df = addDupsGENE(result[[1]], result[[2]], gene)
final.trios.df

}
\seealso{
\code{\link[=findDups]{findDups()}} to find duplicated rows
}
