\name{cna}
\alias{cna}
\docType{data}
\title{
 Copy Number Alteration (CNA) dataset
}
\description{
 A dataset containing the gene names, Entrez ID and CNA values for individuals. It is a subset of a whole genome CNA dataset.
}
\usage{data("cna")}
\format{
  A data frame with 24 observations of 1082 variables.  The first two variables are Hugo_Symbol (for gene name) and Entrez_Gene_Id.  The remaining variables are IDs of individuals.
}
\examples{
data(cna)
}
\keyword{datasets}

\seealso{
\code{\link[=trios]{trios}}
}

