% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrezCNA.R
\name{entrezCNA}
\alias{entrezCNA}
\title{A function to locate gene in CNA data from Gene Expression based on their entrez ID}
\usage{
entrezCNA(cna.row, data, cna, gene)
}
\arguments{
\item{cna.row}{row numbers in trio data matrix with NA for the CNA row column}

\item{data}{a trios data matrix with 4 columns}

\item{cna}{Copy Number Alteration data; genes in rows, individuals in columns}

\item{gene}{Gene Expression data; genes in rows, individuals in columns}
}
\value{
a row number in the CNA data to replace NA in the data
}
\description{
This function uses the trios matrix that we get from the trios() function and fills in the missing values by matching entrez ID in the CNA data to the Gene Expression data.
}
\examples{
#load the example datasets
data(meth)
data(cna)
data(gene)

#Methylation data has multiple genes in one row separated by ";", so we split them
meth.genes = strsplit(as.character(meth$Gene_Symbol), ';')

#get the rows with all NA values
na.meth = removeNA(meth, 5)
na.gene = removeNA(gene, 3)
na.cna = removeNA(cna, 3)

#use the trios function to get the trio data
result = trios("C13orf15", meth, cna, gene, meth.genes, na.meth, na.gene, na.cna)
result

# Extract rows that have missing values
# Since we have one only row in the data here,
# it returns 1, but in case of larger data 
# it will return a vector of row numbers
row = as.integer(which(is.na(result[,3] == TRUE)))

#Now we use the function with the required arguments
val = entrezCNA(row, result, cna, gene)
val

#Then we replace the value in the trio data
result[row,3] = entrezCNA(row, result, cna, gene)
result
}
\seealso{
\link{trios} to get the input data (trios data) used in this function
}
