% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractHumanMethProbeInfo.R
\name{extractHumanMethProbeInfo}
\alias{extractHumanMethProbeInfo}
\title{A function to extract probe information from the human methylation and biomart dataset.}
\usage{
extractHumanMethProbeInfo(
  df,
  modeltype,
  TCGA.meth,
  gene.exp,
  cna,
  trios,
  humanmeth,
  biomart,
  nstartMeth,
  nstartGene,
  nstartCNA,
  type.ind
)
}
\arguments{
\item{df}{A data frame generated from \link{postFilter}}

\item{modeltype}{A string with the name of causal model}

\item{TCGA.meth}{Methylation data matrix; probes in rows, individuals in columns}

\item{gene.exp}{Gene Expression data; genes in rows, individuals in columns}

\item{cna}{Copy Number Alteration data; genes in rows, individuals in columns}

\item{trios}{A trios data matrix with 4 columns}

\item{humanmeth}{The human methylation data matrix with probe ID, gene name, mapinfo, and cpg island information}

\item{biomart}{The biomart data matrix with gene name, gene start and gene end location, and GC content}

\item{nstartMeth}{The column where numeric values begin in methylation data}

\item{nstartGene}{The column where numeric values begin in gene expression data}

\item{nstartCNA}{The column where numeric values begin in copy number alteration data}

\item{type.ind}{A vector of positive or negative ER individuals ID}
}
\value{
A data frame with 19 columns:

IlmnID

\if{html}{\out{<div class="sourceCode">}}\preformatted{ DNA Methylation Probe ID
}\if{html}{\out{</div>}}

UCSC_RefGene_Name

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Target gene name(s) where the Probe ID is found
}\if{html}{\out{</div>}}

Genome_Build

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The human reference genome information
}\if{html}{\out{</div>}}

CHR

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Chromosome containing the CpG
}\if{html}{\out{</div>}}

MAPINFO

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Chromosomal coordinates of the CpG
}\if{html}{\out{</div>}}

UCSC_CpG_Islands_Name

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Chromosomal coordinates of the CpG Island from UCSC.
}\if{html}{\out{</div>}}

Relation_to_UCSC_CpG_Island

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The location of the CpG relative to the CpG island.
}\if{html}{\out{</div>}}

trio_row

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The row number of the gene in the trio data
}\if{html}{\out{</div>}}

Biomart_GeneName

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Since a probe can be found in multiple genes, we look at one gene information at a time.
}\if{html}{\out{</div>}}

Gene.stable.ID

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The identifiers that refer to the same genomic features or species
}\if{html}{\out{</div>}}

Gene.stable.ID.version

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The increments when the set of transcripts linked to a gene changes
}\if{html}{\out{</div>}}

Gene_start

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The location where the gene begins
}\if{html}{\out{</div>}}

Gene_end

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The location where the gene ends
}\if{html}{\out{</div>}}

diff_cpG_mapinfo

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The difference between midpoint of nearby CpG location and the location of methylation probe
}\if{html}{\out{</div>}}

diff_mapinfo_geneStart

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The difference between location of methylation probe and the start of the gene
}\if{html}{\out{</div>}}

gene_length

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The length of the gene (difference between gene start and gene end)
}\if{html}{\out{</div>}}

Methyl_mean

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Mean of the methylation values for the corresponding probe ID
}\if{html}{\out{</div>}}

Methyl_sd

\if{html}{\out{<div class="sourceCode">}}\preformatted{ Standard deviation of the methylation values for the corresponding probe ID
}\if{html}{\out{</div>}}

GC_content

\if{html}{\out{<div class="sourceCode">}}\preformatted{ The content of Guanine-Cytosin for the probe ID.
}\if{html}{\out{</div>}}
}
\description{
This function use the output data from MRGN trio analysis and post filtering to extract additional information on probes for each trio. The extracted details are then used to calculate the difference between CpG in methylation probe and the midpoint of a nearby CpG island, CpG in methylation probe and the start of a gene, the gene length, methylation mean, GC content, and relation to CpG island.
}
\examples{
#Run postFilter() on results from analyzeTrios() to get an object 'df'

\dontrun{
probe_info_data = extractHumanMethProbeInfo(df, "M1.1", meth, gene, cna, final.trios.df, humanmeth, biomart, 5, 3, 3, clinical.pos)
}

}
\seealso{
\link{findTrioAll} \link{addDupsCNA}, \link{addDupsGENE} \link{findPCs} \link{findPCsGeneral} \link{analyzeTrios}
}
