% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDups.R
\name{findDups}
\alias{findDups}
\title{A function to find rows that have duplicated Entrez IDs}
\usage{
findDups(data)
}
\arguments{
\item{data}{A data with genes in rows and individuals in columns}
}
\value{
a vector with row numbers
}
\description{
In genomic datasets, duplicated rows sometimes have either the same gene name and Entrez ID or the same values for all the individuals which introduces repetition in the data. Hence, we extract the duplicated rows so we can remove them later.
}
\examples{
#load the example datasets
data(meth)
data(cna)
data(gene)

#use the function to get the row numbers in the dataset
row = findDups(cna)
row
cna[row,1:5]

}
