% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIndex.R
\name{findIndex}
\alias{findIndex}
\title{A function to track the indices of data matrix}
\usage{
findIndex(data, startCol, GeneNameCol, type.ind, com.ind, type)
}
\arguments{
\item{data}{A data matrix; probes or genes in rows, individuals in columns}

\item{startCol}{The column where numeric values begin in the data}

\item{GeneNameCol}{The column where the gene names are located}

\item{type.ind}{A vector of positive or negative ER individuals ID}

\item{com.ind}{A vector of common individuals between the datasets you want to find PCs for}

\item{type}{A string of individual type ("ER pos" or "ER neg")}
}
\value{
A data matrix with 2 columns; Column 1 is indices in the input data, Column 2 is the indices in data after removing columns with no variance
}
\description{
The data goes through muliple changes before it is used to calculate the PC scores and significantly associated PCs so we keep track of the change through indices in this data matrix.
}
\examples{
#load the datasets
data(clinical.neg)
data(clinical.pos)

#Find common individuals between the methylation and gene expression dataset
com.ind = intersect(colnames(gene)[3:ncol(gene)], colnames(meth)[5:ncol(meth)])

#Use the function to get the indices data matrix
gene.table.pos = findIndex(gene, 3, 1, clinical.pos[,1], com.ind, "Pos")

}
