% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findTrioAll.R
\name{findTrioAll}
\alias{findTrioAll}
\title{A function to form trios between three variables of interest for multiple genes}
\usage{
findTrioAll(meth.data, cna.data, gene.data, nStartMeth, nStartGene, nStartCNA)
}
\arguments{
\item{meth.data}{Methylation data; probes in rows, individuals in columns}

\item{cna.data}{Copy Number Alteration data; genes in rows, individuals in columns}

\item{gene.data}{Gene expression data; genes in rows, individuals in columns}

\item{nStartMeth}{The column where numeric values begin in methylation data}

\item{nStartGene}{The column where numeric values begin in gene expression data}

\item{nStartCNA}{The column where numeric values begin in CNA data}
}
\value{
A data matrix of four columns:

gene.name

\if{html}{\out{<div class="sourceCode">}}\preformatted{      name of the unique gene
}\if{html}{\out{</div>}}

meth.row

\if{html}{\out{<div class="sourceCode">}}\preformatted{      row number of the gene in methylation data
}\if{html}{\out{</div>}}

cna.row

\if{html}{\out{<div class="sourceCode">}}\preformatted{      row number of the gene in cna data
}\if{html}{\out{</div>}}

gene.row

\if{html}{\out{<div class="sourceCode">}}\preformatted{      row number of the gene in gene expression data
}\if{html}{\out{</div>}}
}
\description{
A function to find trios by matching methylation probes, gene expression, and copy number alteration (CNA) for a vector of genes.
}
\examples{
#use the function to form trios of genes between methylation, gene expression, and cna data
trios.df = findTrioAll(meth, cna, gene, 5, 3, 3)
trios.df

}
\seealso{
\link{findDups} to find duplicated rows;
\link{removeNA} to find rows with all NA for numeric values;
\link{trios} to form trio for each gene using gene names;
\link{entrezCNA} to match trios in CNA data with Gene Expression data based on entrez IDs;
\link{entrezGENE} to match trios in Gene Expression data with CNA data based on entrez IDs
}
