\name{gene}
\alias{gene}
\docType{data}
\title{
 Gene Expression dataset
}
\description{
 A dataset containing the gene names, Entrez ID and expression values for individuals. It is a subset of a whole genome gene expression dataset.
}
\usage{data("gene")}
\format{
  A data frame with 22 observations of 1102 variables.  The first two variables are Hugo_Symbol (for gene name) and Entrez_Gene_Id.  The remaining variables are IDs of individuals.
}
\examples{
data(gene)
}
\keyword{datasets}

\seealso{
\code{\link[=trios]{trios}}
}

