% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noVar.R
\name{noVar}
\alias{noVar}
\title{A function to find columns in a data that have no variance}
\usage{
noVar(data)
}
\arguments{
\item{data}{A dataset with numeric values; individuals in rows and probes or genes in column}
}
\value{
A vector of column numbers that either have 0 variance or 2 or less than 2 non-NA values
}
\description{
Data with no variance might be easy to predict but has no uncertainity so we do not include those columns in our analysis.
}
\examples{
# Transpose the data since individuals are in columns
data = t(meth[,5:ncol(meth)])

# Returns column numbers
col = noVar(data)
col
data[,col]
}
