% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postFilter.R
\name{postFilter}
\alias{postFilter}
\title{A function apply marginal test on the MRGN inferred models}
\usage{
postFilter(res)
}
\arguments{
\item{res}{A data frame of 16 columns; the object generated from \link{analyzeTrios}:}
}
\value{
The input data frame with the updated inferred models column
}
\description{
Since MRGN only applies conditional test with the inferred models, it does not account for the marginal test that could influence the results. So, we use this function to update the inferred models.
}
\examples{
#Use the function to perform the MRGN inference with confounding variables

\dontrun{
df = postFilter(final.result)
}

}
