% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNA.R
\name{removeNA}
\alias{removeNA}
\title{A function to find rows that have all NA values in a data}
\usage{
removeNA(data, nStart)
}
\arguments{
\item{data}{A data with genes in rows and individuals in columns}

\item{nStart}{The column where numeric values begin}
}
\value{
a vector with row numbers
}
\description{
This function uses a data with genes in rows and individuals in columns to extract row numbers that have all NA values
}
\examples{
#load the example datasets
data(meth)

#use the function and get the row numbers in the dataset that have NAs
removeNA(meth, 5)
}
