\name{sig.pcs.gene}
\alias{sig.pcs.gene}
\docType{data}
\title{
  Significantly associated PCs for ER + individuals in gene expression dataset
}
\description{
  A data with the PCs that were significantly associated with each individual trio after the Principal component score matrix was calculated for the gene expression dataset.
}
\usage{data("sig.pcs.gene")}
\format{
  The format is:
List of 22
 $ : Named int [1:10] 1 2 3 4 5 6 8 10 12 572
  ..- attr(*, "names")= chr [1:10] "V1" "V2" "V3" "V4" ...
 $ : Named int [1:14] 1 3 4 5 6 11 13 14 19 20 ...
  ..- attr(*, "names")= chr [1:14] "V1" "V3" "V4" "V5" ...
 $ : Named int [1:11] 2 3 4 5 6 7 8 9 10 12 ...
  ..- attr(*, "names")= chr [1:11] "V2" "V3" "V4" "V5" ...
 $ : Named int [1:13] 2 3 4 7 8 9 10 12 13 19 ...
  ..- attr(*, "names")= chr [1:13] "V2" "V3" "V4" "V7" ...
 $ : Named int [1:14] 1 2 3 5 6 7 8 10 13 16 ...
  ..- attr(*, "names")= chr [1:14] "V1" "V2" "V3" "V5" ...
 $ : Named int 1
  ..- attr(*, "names")= chr "V1"
 $ : Named int [1:18] 1 3 4 5 6 7 11 12 13 14 ...
  ..- attr(*, "names")= chr [1:18] "V1" "V3" "V4" "V5" ...
 $ : Named int [1:10] 1 2 3 4 5 6 7 9 10 11
  ..- attr(*, "names")= chr [1:10] "V1" "V2" "V3" "V4" ...
 $ : Named int [1:11] 1 2 3 4 7 9 14 22 26 27 ...
  ..- attr(*, "names")= chr [1:11] "V1" "V2" "V3" "V4" ...
 $ : Named int [1:14] 1 3 4 5 6 7 8 9 12 13 ...
  ..- attr(*, "names")= chr [1:14] "V1" "V3" "V4" "V5" ...
 $ : Named int [1:11] 1 2 4 5 6 7 8 10 14 41 ...
  ..- attr(*, "names")= chr [1:11] "V1" "V2" "V4" "V5" ...
 $ : Named int [1:5] 3 4 8 9 24
  ..- attr(*, "names")= chr [1:5] "V3" "V4" "V8" "V9" ...
 $ : Named int [1:13] 1 2 3 5 6 7 9 11 21 24 ...
  ..- attr(*, "names")= chr [1:13] "V1" "V2" "V3" "V5" ...
 $ : Named int [1:18] 1 3 4 5 6 7 8 9 10 12 ...
  ..- attr(*, "names")= chr [1:18] "V1" "V3" "V4" "V5" ...
 $ : Named int [1:16] 1 2 3 4 5 6 7 10 12 14 ...
  ..- attr(*, "names")= chr [1:16] "V1" "V2" "V3" "V4" ...
 $ : Named int [1:6] 1 2 5 8 13 572
  ..- attr(*, "names")= chr [1:6] "V1" "V2" "V5" "V8" ...
 $ : Named int [1:13] 1 2 3 5 8 11 12 13 16 19 ...
  ..- attr(*, "names")= chr [1:13] "V1" "V2" "V3" "V5" ...
 $ : Named int [1:2] 2 9
  ..- attr(*, "names")= chr [1:2] "V2" "V9"
 $ : Named int [1:7] 1 2 3 5 6 12 17
  ..- attr(*, "names")= chr [1:7] "V1" "V2" "V3" "V5" ...
 $ : Named int [1:12] 1 2 3 4 6 7 8 9 10 13 ...
  ..- attr(*, "names")= chr [1:12] "V1" "V2" "V3" "V4" ...
 $ : Named int [1:14] 1 2 3 4 5 6 9 10 13 20 ...
  ..- attr(*, "names")= chr [1:14] "V1" "V2" "V3" "V4" ...
 $ : Named int [1:11] 1 2 3 4 6 7 9 18 28 35 ...
  ..- attr(*, "names")= chr [1:11] "V1" "V2" "V3" "V4" ...
}
\examples{
data(sig.pcs.gene)
}

\keyword{datasets}

\seealso{
\code{\link[=findPCs]{findPCs()}}
}
