% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trios.R
\name{trios}
\alias{trios}
\title{A function to form trios between three variables of interest for a gene}
\usage{
trios(gene.name, meth, cna, gene, meth.genes, na.meth, na.gene, na.cna)
}
\arguments{
\item{gene.name}{A unique gene name}

\item{meth}{Methylation data matrix; probes in rows, individuals in columns}

\item{cna}{Copy Number Alteration data; genes in rows, individuals in columns}

\item{gene}{Gene Expression data; genes in rows, individuals in columns}

\item{meth.genes}{List of all genes in the methylation data}

\item{na.meth}{a vector of rows in methylation data that are all NAs}

\item{na.gene}{a vector of rows in gene expression data that are all NAs}

\item{na.cna}{a vector of rows in CNA data that are all NAs}
}
\value{
a row vector of length 4:

gene.name

\if{html}{\out{<div class="sourceCode">}}\preformatted{      name of the unique gene
}\if{html}{\out{</div>}}

meth.row

\if{html}{\out{<div class="sourceCode">}}\preformatted{      row number of the gene in methylation data
}\if{html}{\out{</div>}}

cna.row

\if{html}{\out{<div class="sourceCode">}}\preformatted{      row number of the gene in cna data
}\if{html}{\out{</div>}}

gene.row

\if{html}{\out{<div class="sourceCode">}}\preformatted{      row number of the gene in gene expression data
}\if{html}{\out{</div>}}
}
\description{
A function to find trios by matching methylation probes, gene expression, and copy number alteration (CNA) for an individual gene.
}
\examples{
#load the example datasets
data(meth)
data(cna)
data(gene)

#Methylation data has multiple genes in one row separated by ";", so we split them
meth.genes = strsplit(as.character(meth$Gene_Symbol), ';')

#get the rows with all NA values
na.meth = removeNA(meth, 5)
na.gene = removeNA(gene, 3)
na.cna = removeNA(cna, 3)

#Now we use the function with the required arguments

#Example 1

result = trios("C3orf35", meth, cna, gene, meth.genes, na.meth, na.gene, na.cna)
result

# Example 2

# 'result' below should have NA in the gene.row column 
# because no match is found.
result = trios("OR4P4", meth, cna, gene, meth.genes, na.meth, na.gene, na.cna)
result

# Example 3

# 'result' below should to be NULL 
# since this gene has all NA values in the methylation dataset
result = trios("RBL2", meth, cna, gene, meth.genes, na.meth, na.gene, na.cna)
result

}
