% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findIndex.R
\name{findIndex}
\alias{findIndex}
\title{A function tracking index changes in data matrix after data cleaning}
\usage{
findIndex(data, startCol, GeneNameCol, type.ind, com.ind)
}
\arguments{
\item{data}{A data matrix with probes or genes in rows and individuals in columns.}

\item{startCol}{The column where numeric values begin in the data.}

\item{GeneNameCol}{The column where the gene names are located.}

\item{type.ind}{A vector of positive or negative ER individuals ID.  If there are no subtypes, use the same argument for 'com.ind' here.}

\item{com.ind}{A vector of common individuals shared across the datasets for which principal component analyses are performed.}
}
\value{
A data matrix with 2 columns; Column 1 is indices in the input data, Column 2 is the indices in data after removing columns with no variance
}
\description{
After finding the individuals of interest in data, retain rows (probes or genes) of data that have less than three NAs.  Output the index of the row before and after this processing.
}
\examples{
#load the datasets
data(gene)
data(meth)
data(clinical.neg)
data(clinical.pos)

#Find common individuals between the methylation and gene expression dataset
com.ind = intersect(colnames(gene)[3:ncol(gene)], colnames(meth)[5:ncol(meth)])

# Use the function to get the indices data matrix
gene.table.pos = findIndex(data = gene, startCol = 3, GeneNameCol = 1, 
        type.ind = clinical.pos[,1], com.ind = com.ind)

}
